# api/v1/static/swagger.yaml
openapi: 3.0.0
info:
  title: JWS COLLECTIONS API
  version: 1.0.0
  description: API for managing orders, products, users, and checkout in an e-commerce system.
servers:
  - url: /api/v1
    description: Main API server
paths:
...
  /products:
    get:
      summary: Get all products
      description: Returns a JSON format of all products with optional filtering and pagination.
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Number of products per page
          required: false
          schema:
            type: integer
            default: 10
        - name: category
          in: query
          description: Filter products by category
          required: false
          schema:
            type: string
        - name: brand
          in: query
          description: Filter products by brand
          required: false
          schema:
            type: string
        - name: min_price
          in: query
          description: Filter products with a minimum price
          required: false
          schema:
            type: number
            format: float
        - name: max_price
          in: query
          description: Filter products with a maximum price
          required: false
          schema:
            type: number
            format: float
        - name: color
          in: query
          description: Filter products by color
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of products with pagination details
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  per_page:
                    type: integer
                  total_products:
                    type: integer
                  total_pages:
                    type: integer
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
        '404':
          description: No products available
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  per_page:
                    type: integer
                  total_products:
                    type: integer
                  total_pages:
                    type: integer
                  products:
                    type: array
                    items:
                      type: object

    post:
      summary: Create a new product
      description: Creates a new product in the catalog. Admin access required.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shoe_name:
                  type: string
                  example: "Air Max 95"
                shoe_category:
                  type: string
                  example: "sneakers"
                shoe_brand:
                  type: string
                  example: "Nike"
                shoe_price:
                  type: float
                  example: 10.0
                shoe_color:
                  type: string
                  example: "brown"
                shoe_image:
                  type: string
                  example: "https://www.imgur.com/..."
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "20303dw.."
                  shoe_image:
                    type: string
                    example: "https://images.puma.com/.."                  
                  shoe_price:
                    type: float
                    example: 121
                  created_at:
                    type: string
                    example: "2024-08-17T20:04:01.063292"
                  shoe_brand:
                    type: string
                    example: "Puma"
                  shoe_category:
                    type: string
                    example: "sneakers"
                  shoe_color:
                    type: string
                    example: "yellow"
                  shoe_name:
                    type: string
                    example: "Wild ..."
                  updated_at:
                    type: string
                    example: "2024-08-17T20:04:01.063292"
        '400':
          description: Bad request or invalid data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Invalid data'
  /products/{id}:
    get:
      summary: Get a product by ID
      description: Returns product details based on the provided ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the product to retrieve
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Product not found'

    put:
      summary: Update a product
      description: Updates product details based on the provided ID. Admin access required.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the product to retrieve
      security:
        - bearerAuth: []
        - role: ['admin']
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                shoe_name:
                  type: string
                shoe_category:
                  type: string
                shoe_brand:
                  type: string
                shoe_price:
                  type: number
                  format: float
                shoe_color:
                  type: string
        required: true
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "20303dw.."
                  shoe_image:
                    type: string
                    example: "https://images.puma.com/.."                  
                  shoe_price:
                    type: float
                    example: 121
                  created_at:
                    type: string
                    example: "2024-08-17T20:04:01.063292"
                  shoe_brand:
                    type: string
                    example: "Puma"
                  shoe_category:
                    type: string
                    example: "sneakers"
                  shoe_color:
                    type: string
                    example: "yellow"
                  shoe_name:
                    type: string
                    example: "Wild ..."
                  updated_at:
                    type: string
                    example: "2024-08-17T20:04:01.063292"
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Product not found'

    delete:
      summary: Delete a product
      description: Deletes a product based on the provided ID. Admin access required.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the product to retrieve
      security:
        - bearerAuth: []
        - role: ['admin']
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Product deleted successfully'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Product not found'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT